+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix bug with useServiceWorkerUX.js when not running the service worker (i.e. in dev mode)
 - Fix AssetPageTitleWithGroupsWidgetDecorator plugin to not break when group assets/relationships are disabled
+- Fix MoveAssetActionProvider plugin to filter the current asset out of possible move destinations
 
 ## [0.9.0] - 2023-02-02